+++ src/index.ts
@@ -115,7 +115,7 @@ export { PaletteModule,
          CasesService,
          ActivityModule,
          CaseHeaderModule,
-         EventTriggerModule,
+         EventManagementModule,
          EventLogModule,
          SearchFiltersModule,
          CaseResolver,
