+++ src/shared/components/event-management/components/task-assigned/task-assigned.component.spec.ts
