+++ src/shared/components/event-management/event-trigger/event-trigger.component.html
