+++ src/shared/components/event-management/index.ts
@@ -1,0 +1,8 @@
+export * from './event-start/event-start.component';
+export * from './event-trigger/event-trigger.component';
+export * from './components/multiple-tasks-exist/multiple-tasks-exist.component';
+export * from './components/no-tasks-available/no-tasks-available.component';
+export * from './components/task-assigned/task-assigned.component';
+export * from './components/task-cancelled/task-cancelled.component';
+export * from './components/task-conflict/task-conflict.component';
+export * from './event-management.module';
