--- src/shared/components/event-trigger/index.ts
@@ -1,2 +1,0 @@
-export * from './event-trigger.component';
-export * from './event-trigger.module';
